Make it macOS compatible.

+++ install.sh
@@ -8,6 +8,8 @@ if [[ ! -e ~/.zplug ]]; then
   curl -sL --proto-redir -all,https https://raw.githubusercontent.com/zplug/installer/master/installer.zsh| zsh
 fi
 
+UNAME=$(uname)
+
 for f in $(ls -A dotfiles)
 do
   if [ -L ~/$f ]; then
@@ -18,29 +20,31 @@ do
   fi
 
   FILE_PATH="$PWD/dotfiles/$f"
-  if [ -d $FILE_PATH ]; then
+  if [ -d $FILE_PATH ] && [ ! $UNAME == 'Darwin' ]; then
     ln -sd $FILE_PATH ~/$f
-  elif [ -f $FILE_PATH ]; then
+  else
     ln -s $FILE_PATH ~/$f
   fi
 done
 
-for f in $(ls -A config)
-do
-  if [ -L ~/.config/$f ]; then
-    rm ~/.config/$f
-  fi
-  if [ -e ~/.config/$f ]; then
-    mv ~/.config/$f ~/.config/$f.bak
-  fi
+if [ ! $UNAME == 'Darwin' ]; then
+  for f in $(ls -A config)
+  do
+    if [ -L ~/.config/$f ]; then
+      rm ~/.config/$f
+    fi
+    if [ -e ~/.config/$f ]; then
+      mv ~/.config/$f ~/.config/$f.bak
+    fi
 
-  FILE_PATH="$PWD/config/$f"
-  if [ -d $FILE_PATH ]; then
-    ln -sd $FILE_PATH ~/.config/$f
-  elif [ -f $FILE_PATH ]; then
-    ln -s $FILE_PATH ~/.config/$f
-  fi
-done
+    FILE_PATH="$PWD/config/$f"
+    if [ -d $FILE_PATH ]; then
+      ln -sd $FILE_PATH ~/.config/$f
+    else
+      ln -s $FILE_PATH ~/.config/$f
+    fi
+  done
+fi
 
 if [ ! -e ~/.vim/bundle/Vundle.vim ]; then
   git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
